+++ packages/compiled-assets/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prairielearn/compiled-assets
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [638d27585]
+  - @prairielearn/html@3.0.0
+
 ## 1.0.1
 
 ### Patch Changes