+++ packages/named-locks/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @prairielearn/named-locks
 
+## 1.3.1
+
+### Patch Changes
+
+- Updated dependencies [00d1b045d]
+  - @prairielearn/postgres@1.6.1
+
+## 1.3.0
+
+### Minor Changes
+
+- 4cc962358: Add `tryWithLock` function
+
+## 1.2.0
+
+### Minor Changes
+
+- 400a0b901: Allow locks to be automatically renewed
+
 ## 1.1.0
 
 ### Minor Changes